[dhcp_probe] Add a Wants= on network-online.target
According to https://systemd.io/NETWORK_ONLINE/ network-online.target is
an "active target". This means that this target will not be started unless
explicitely needed by another service

We already have an After=network-online.target but After= does not start
the target, it only guarantees ordering

This means that if dhcp_probe is the only network-related service,network-
online.target will not be started and dhcp_probe will be started too early
and fail


See the first FAQ in the page documented above.