Limit depth of searches for playbook directories
This patch fixes the way DebOps looks for playbook directories in
Ansible Collections. Previously, unrelated 'playbooks/' subdirectories
under 'tests/integration/' were considered as valid playbook directories
as well. Now the search for the 'playbooks/' subdirectory in
a collection is limited to specific level of subdirectories, which
should avoid the mistake.