Clarify the use of the '.env' files
The '.env' file at the root of the project directory is meant for local
environment variables not tracked by the version control.

The 'ansible/views/*/.env' file cannot be read at the start of the
script before setting up the configuration tree because at that point we
don't know if we even have any views present. So let's drop its specific
entry in the '.gitignore' file so that there's no special meaning about
that file.