Remove '/debops' from .gitignore
The 'debops/' directory or symlink was an old way to select which DebOps
version was used in the project via a symlink to the monorepo. Since we
are now using Ansible Collections, there's no need to symlink the
monorepo at the root of the project anymore, and there's no support for
that in the scripts as well. Let's remove a reference to that to avoid
confusion.