[ci] Generate ansible.cfg to fix linting issues
This patch ensures that ansible-lint can find both DebOps collections
included in the monorepo as well as dependent collections installed in
the '~/.cache/ansible-compat/` path. To do this reliably, we create
a stub 'ansible.cfg' file in the root of the repository which tells
ansible-lint where it can find Ansible Collections.

This needs to be generated dynamically because 'ansible-compat' uses
a hashed path in the '~/.cache/' directory to store downloaded
connections.

Since ansible-playbook will also use the generated 'ansible.cfg' during
syntax check, we need to install the dependent collections manually for
this test to work.

Ref: https://github.com/ansible/ansible-lint/issues/3344