[ifupdown]: escape and unescape interface names in systemd unit name
This fixes systemd unit names escaping. For one it let us escape the "hyphen"
in the network interface name when we add them to systemd unit name in ifupdown
iface@ services.
This restores unescaping of the systemd instance name to get the network interface
name (the removal of this unescaping was a hack to avoid escaping the systemd unit
instance name).

Closes #2371.

(cherry picked from commit 8447e5938d499722c9761545e29da0854fd6ef89)