[slapd] Partial update for Bookworm
The ppolicy schema is now internal to the slapo-ppolicy(5) module,
meaning that:

  a) It can't be loaded (harmless, the role already checks for
     "already exists" error messages); and

  b) Updating an existing slapd installation from pre-Bookworm to
     Bookworm will cause pain, see [1] and [2].

Not sure if that should be documented somewhere...

Also, the fusiondirectory packages aren't in Bookworm [3], that's sure to
cause pain. Perhaps we should import the schema files into the repo instead
(GPL2+ according to the copyright file in the Debian package).

[1] /usr/share/doc/slapd/README.Debian.gz)
[2] https://www.debian.org/releases/bookworm/amd64/release-notes/ch-information.en.html#openldap-2.5
[3] https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=1016029