[dovecot] Add ability to list users using doveadm
Before this change, dovecot could only check if users existed, but it
could not iterate over them. That is, the following would not work

    doveadm user '*'

Nor any command that takes `-A` (all-users), like so:

    doveadm purge -A

With this change, a new default query is added that works like the
existing default queries. Instead of searching for a user, it lists
them however.

(cherry picked from commit 15f8edf3621160004c2857137a0e889621413508)