Update installation documentation with pipx instead of pip
As pointed by the issue #2410, the documentation is not applicable for Debian
bookworm which implemented PEP668.

This commit provides simple examples to use pipx in place of pip to still
satisfy the first method of installation :

"As a Python package in a Python environment of a given UNIX account.
  Different UNIX accounts can use their own versions of DebOps."

(cherry picked from commit 6ca35cdca2160243f6799405a7a67fbd640dc032)