Add separate option enabling to set basedir variable
This is necessary in a case, when someone needs to have a custom
value for this variable to be able to define custom reprepro repository
structure for a repository.

It follows the same approach that is defined for the "outdir" variable,
that can be set separately, if a defuault generated value can't be
used.

(cherry picked from commit 71ad563a20a0391db368f35d0f6c795fc4d696e5)