[root_account] Explain issue with 'passlib' module
If the 'passlib' Python module is not installed and available for
Ansible, the 'root_account' role (as the first of multiple roles) might
error out without specific explanation, and 'debops__no_log' variable
might not work correctly in this case to help explain the error.

This commit adds a comment which explains the issue, hopefully helping
resolve it in future cases.