Change the way DebOps checks current user
The 'os.getlogin()' function relies on the glibc 'getlogin()' function,
which requires a control terminal (tty, pty) to work. In an environment
without the control terminal (systemd service, remote execution) it
results in the 'OSError: [Errno 6] No such device or address' exception.

Switching to 'getpass.getuser()' function is a recommended alternative.