fix(netbox): match upstream for SESSION_FILE_PATH (`None`, not `False`)
Setting `SESSION_FILE_PATH` to `False` has worked for many years but it is
not good to derive from upstream config example. We now either set it to
`None` or to a string.

Those lines where changed in 92cc8852982fa9e74942ff3e3113d4dbdfdf1849
(#2435) and thus I dug deeper. But the derivation from upstream (`False`
and not `None`) was there before. With #2435, it was just changed to ""
(empty string). Both `False` and empty string are not ideal I guess (see
`settings.py`).

Ref: https://github.com/netbox-community/netbox/blob/27d15615b3590000901984bddfa7a815cd1343d8/netbox/netbox/configuration_example.py#L222
Ref: https://github.com/netbox-community/netbox/blob/27d15615b3590000901984bddfa7a815cd1343d8/netbox/netbox/settings.py#L348-L349