silence shellcheck with embedded shell script have jinja templating
shellcheck does not support jinja templating
https://github.com/koalaman/shellcheck/issues/1306

Note that lib/tests/check-shell already has code to ignore these files,
but I believe it is better to silence the error in the shell script
templates than to maintain a hardcoded list of template files to ignore.

Note that shellcheck disable=all is for shellcheck >= 0.8, ie starting
from bookworm.