[logrotate] Add support for systemd timer
Since Debian Bookworm, logrotate is not implemented via cron but with a systemd
timer.  On previous systems and Ubuntu, the crontab and the systemd timer are
both installed.

While is it not harmful to have multiple schedulers calling logrotate, it is
unnecessary, and it also pull cron as a dependency on systems where it may not
be installed.

This commit shift from cron to a systemd timer starting at Debian>11 and
Ubuntu>24.